Add document type webhook events and use fluent builders for content and content types
