+++ src/Umbraco.Core/Webhooks/Events/ContentType/DocumentTypeMovedWebhookEvent.cs
@@ -1,0 +1,25 @@
+using Microsoft.Extensions.Options;
+using Umbraco.Cms.Core.Configuration.Models;
+using Umbraco.Cms.Core.Notifications;
+using Umbraco.Cms.Core.Services;
+using Umbraco.Cms.Core.Sync;
+
+namespace Umbraco.Cms.Core.Webhooks.Events.ContentType;
+
+[WebhookEvent("Document Type Moved")]
+public class DocumentTypeMovedWebhookEvent : WebhookEventBase<ContentTypeMovedNotification>
+{
+    public DocumentTypeMovedWebhookEvent(
+        IWebhookFiringService webhookFiringService,
+        IWebhookService webHookService,
+        IOptionsMonitor<WebhookSettings> webhookSettings,
+        IServerRoleAccessor serverRoleAccessor)
+        : base(webhookFiringService, webHookService, webhookSettings, serverRoleAccessor)
+    {
+    }
+
+    public override string Alias => "documentTypeMoved";
+
+    public override object? ConvertNotificationToRequestPayload(ContentTypeMovedNotification notification)
+        => notification.MoveInfoCollection;
+}
