+++ src/Umbraco.Core/Webhooks/Events/ContentType/DocumentTypeSavedWebhookEvent.cs
@@ -1,0 +1,25 @@
+using Microsoft.Extensions.Options;
+using Umbraco.Cms.Core.Configuration.Models;
+using Umbraco.Cms.Core.Notifications;
+using Umbraco.Cms.Core.Services;
+using Umbraco.Cms.Core.Sync;
+
+namespace Umbraco.Cms.Core.Webhooks.Events.ContentType;
+
+[WebhookEvent("Document Type Saved")]
+public class DocumentTypeSavedWebhookEvent : WebhookEventBase<ContentTypeSavedNotification>
+{
+    public DocumentTypeSavedWebhookEvent(
+        IWebhookFiringService webhookFiringService,
+        IWebhookService webHookService,
+        IOptionsMonitor<WebhookSettings> webhookSettings,
+        IServerRoleAccessor serverRoleAccessor)
+        : base(webhookFiringService, webHookService, webhookSettings, serverRoleAccessor)
+    {
+    }
+
+    public override string Alias => "documentTypeSaved";
+
+    public override object? ConvertNotificationToRequestPayload(ContentTypeSavedNotification notification)
+        => notification.SavedEntities;
+}
