+++ src/Umbraco.Core/Webhooks/Events/ContentType/MediaTypeChangedWebhookEvent.cs
@@ -1,0 +1,25 @@
+using Microsoft.Extensions.Options;
+using Umbraco.Cms.Core.Configuration.Models;
+using Umbraco.Cms.Core.Notifications;
+using Umbraco.Cms.Core.Services;
+using Umbraco.Cms.Core.Sync;
+
+namespace Umbraco.Cms.Core.Webhooks.Events.ContentType;
+
+[WebhookEvent("Media Type Changed")]
+public class MediaTypeChangedWebhookEvent : WebhookEventBase<MediaTypeChangedNotification>
+{
+    public MediaTypeChangedWebhookEvent(
+        IWebhookFiringService webhookFiringService,
+        IWebhookService webHookService,
+        IOptionsMonitor<WebhookSettings> webhookSettings,
+        IServerRoleAccessor serverRoleAccessor)
+        : base(webhookFiringService, webHookService, webhookSettings, serverRoleAccessor)
+    {
+    }
+
+    public override string Alias => "mediaTypeChanged";
+
+    public override object? ConvertNotificationToRequestPayload(MediaTypeChangedNotification notification)
+        => notification.Changes;
+}
