+++ src/Umbraco.Core/Webhooks/Events/ContentType/MemberTypeChangedWebhookEvent.cs
@@ -1,0 +1,25 @@
+using Microsoft.Extensions.Options;
+using Umbraco.Cms.Core.Configuration.Models;
+using Umbraco.Cms.Core.Notifications;
+using Umbraco.Cms.Core.Services;
+using Umbraco.Cms.Core.Sync;
+
+namespace Umbraco.Cms.Core.Webhooks.Events.ContentType;
+
+[WebhookEvent("Member Type Changed")]
+public class MemberTypeChangedWebhookEvent : WebhookEventBase<MemberTypeChangedNotification>
+{
+    public MemberTypeChangedWebhookEvent(
+        IWebhookFiringService webhookFiringService,
+        IWebhookService webHookService,
+        IOptionsMonitor<WebhookSettings> webhookSettings,
+        IServerRoleAccessor serverRoleAccessor)
+        : base(webhookFiringService, webHookService, webhookSettings, serverRoleAccessor)
+    {
+    }
+
+    public override string Alias => "memberTypeChanged";
+
+    public override object? ConvertNotificationToRequestPayload(MemberTypeChangedNotification notification)
+        => notification.Changes;
+}
